🐛 `sparse`: fix `_spbase.count_nonzero` return type

+++ scipy-stubs/sparse/_lil.pyi
@@ -147,9 +147,9 @@ class _lil_base(_spbase[_SCT, tuple[int, int]], IndexMixin[_SCT, tuple[int, int]
     @override
     def resize(self, /, *shape: int) -> None: ...  # pyright: ignore[reportIncompatibleMethodOverride]
 
-    # NOTE: Adding `@override` here will crash stubtest (mypy 1.13.0)
+    # NOTE: Adding `@override` here will crash stubtest (mypy 1.15.0)
     @overload
-    def count_nonzero(self, /, axis: None = None) -> int: ...
+    def count_nonzero(self, /, axis: None = None) -> np.intp: ...
     @overload
     def count_nonzero(self, /, axis: op.CanIndex) -> onp.Array1D[np.intp]: ...
